requirements.txt
serpapi
python-dateutil
newspaper4k
openai
python-dotenv
bertopic
streamlit
plotly
scikit-learn
pandas
voyageai
numpy
odfpy

